requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
lxml-html-clean==0.4.1
